refactor(ui+core): address code-review findings from PR #1245
Driven by an external Codex code review; all reasonably-scoped
findings addressed in-place rather than punted to follow-ups.

Must-fix:
- App.tsx onboarding completion was still pushing the removed
  `/b/<board>/<session>/` shape on success — would have hit the
  fallback route and silently failed to open the new session. Now uses
  the centralized `sessionPath` / `boardPath` builders.

Worth-fixing-now:
- SessionRepository was still LEFT JOIN-ing `boards` and threading a
  `boardSlug` arg through `rowToSession` even though flat `/s/<short>/`
  URLs don't need a slug. Dropped the join and the param at all 8 call
  sites; `worktrees` join stays for `Session.worktree_board_id`.
- `Worktree.url` / `Artifact.url` were emitted unconditionally; a
  `/w/<short>/` link for an unplaced worktree would resolve the
  entity but have nowhere to switch the canvas to. Now gated on
  `row.board_id`. Same gating applied to `Session.url`.
- Type-field doc comments for `Worktree.url` / `Artifact.url` /
  `Session.url` / `Board.url` were describing the pre-flatten format
  (e.g. `{baseUrl}/b/{boardId}/w/{...}/`). Updated to the actual
  emitted shape (`{baseUrl}/ui/<seg>/<short>/`).
- The 50ms deferred recenter in useUrlState had no staleness guard —
  rapid URL changes could fire a stale recenter after a newer
  navigation. Stored in a ref so each transition clears the previous
  pending timer; unmount cleanup added.
- Route declarations in outer App.tsx hardcoded segment letters
  (`/b/`, `/s/`, `/w/`, `/a/`). Now built from the shared
  `ENTITY_PATH_SEGMENTS` const so route table and URL builders can't
  drift.
- Pure URL-param → entity-ID resolvers (`resolveByShortIdPure`,
  `resolveBoardFromUrlPure`, etc.) moved out of the hook file into
  `apps/agor-ui/src/utils/urlResolution.ts`. Zero hook deps; living
  next to the hook just made them harder to test/reuse. Test imports
  updated.

Nice-to-haves:
- `fullUrl()` in core strips a trailing slash off `baseUrl` so a
  misconfigured `daemon.public_url` doesn't produce double-slashed
  URLs.
- `goToWorktree` / `goToArtifact` now push-first and only fall back
  to a direct recenter when the URL didn't actually change — mirrors
  `goToSession` and avoids a prior double-recenter animation.
- Stale `/b/*` comment in outer App.tsx updated. Unused
  `{urlBoardParam, urlSessionShortId, buildUrl}` return from
  `useUrlState` dropped (no consumer).

Net-positive small refactor absorbed:
- `ArtifactRepository.resolveId` was hand-rolling the short-ID
  prefix-match pattern with an unbounded SELECT; now goes through the
  centralized `resolveByShortIdPrefix` + `RESOLVE_SHORT_ID_FETCH_LIMIT`
  helper that the other repos already use.

Considered + rejected (logged for the record):
- A `useLatestRef<T>` helper to dedupe the `useRef(value);
  ref.current = value` pattern. Tried it, but biome's
  `useExhaustiveDependencies` heuristic only recognizes refs created
  via `useRef(...)` directly, so the helper triggered noisy
  false-positives. Left both call sites inline with a comment
  explaining why.
- Changing artifact React Flow node IDs from `board_object.object_id`
  to `artifact_id`: would require reworking board-object mutation
  paths. The `data.artifactId` fallback scan in
  `SessionCanvas.recenterOnNode` is the right local trade-off.
